FIxed Hero Mobile styles. Removed useStyles in Dropdown Search
