+++ apps/climatemappedafrica/src/components/DropdownSearch/DropdownSearch.snap.js
@@ -1,0 +1,46 @@
+// Jest Snapshot v1, https://goo.gl/fbAQLP
+
+exports[`<DropdownSearch /> renders unchanged 1`] = `
+<div>
+  <div
+    class="MuiBox-root css-0"
+    id="location-search"
+  >
+    <p
+      class="MuiTypography-root MuiTypography-body1 css-1u0kiq2-MuiTypography-root"
+    >
+      Search for a location
+    </p>
+    <div
+      class="MuiInputBase-root MuiInputBase-colorPrimary MuiInputBase-adornedEnd css-78q71b-MuiInputBase-root"
+    >
+      <input
+        aria-label="search"
+        class="MuiInputBase-input MuiInputBase-inputAdornedEnd css-yz9k0d-MuiInputBase-input"
+        placeholder="Search for a location"
+        type="text"
+        value=""
+      />
+      <button
+        class="MuiButtonBase-root MuiIconButton-root MuiIconButton-colorPrimary MuiIconButton-sizeSmall css-oru44q-MuiButtonBase-root-MuiIconButton-root"
+        root="[object Object]"
+        tabindex="0"
+        type="button"
+      >
+        <svg
+          aria-hidden="true"
+          class="MuiSvgIcon-root MuiSvgIcon-fontSizeMedium css-kzlhbf-MuiSvgIcon-root"
+          focusable="false"
+          viewBox="0 0 48 48"
+        />
+        <span
+          class="MuiTouchRipple-root css-8je8zh-MuiTouchRipple-root"
+        />
+      </button>
+    </div>
+    <div
+      class="MuiBox-root css-79elbk"
+    />
+  </div>
+</div>
+`;
